pyproject.toml
[tool.poetry]
name = "specup"
version = "0.1.0"
description = ""
authors = ["Luca <luca@bluewhale.kr>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.0"
flask-classful = "^0.14.1"
marshmallow = "^2.18"
flask-marshmallow = "^0.9.0"
marshmallow-sqlalchemy="^0.15.0"
sqlalchemy = "^1.2"
alembic = "^1.0"
flask-sqlalchemy = "^2.3"
flask-login = "^0.4.1"
sqlalchemy-utils = "^0.33.9"
passlib = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-flask="^0.14.0"
pymysql="^0.9.3"
pytest-cov="^2.0.0"
factory_boy="^2.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
